To puto

+++ code/src/tempo.cpp
@@ -18,14 +18,14 @@ void Tempo::iniciarTempo()
     this->tempoInicio = SDL_GetTicks();
 }
 
-long int Tempo::pegarTempoCorrido()
+int Tempo::pegarTempoCorrido()
 {
     return SDL_GetTicks() - this->tempoInicio;
 }
 
 void Tempo::atrasarTempo()
 {
-    long int atraso = this->pegarTempoCorrido();
+    int atraso = this->pegarTempoCorrido();
     
     if((1000 / Tela::FPS) > atraso)
         SDL_Delay((1000 / Tela::FPS) - atraso);
